NodePort, LoadBalancer and ClusterIP from k8s Node support for AntreaProxy on Linux
This PR implements:
 - The connection request of NodePort whose client is from remote or
local host.
 - The connection request of LoadBalancer whose client is from remote
or localhost.
 - The connection request of ClusterIP whose client is from localhost.

For NodePort support, an ipset is used to store NodePort IP and port
in an entry IP:port,protocol. The iptables rules are used to match the
ipset as destination and do DNAT with a virtual IP. A routing entry
is used to route the packets whose destination is the virtual IP to
Antrea gateway.

For LoadBalancer support, a routing entry will be created for every
ingress IP to route the packets from remote or localhost to Antrea
gateway.

For ClusterIP support, a routing entry is always used to route the
packets to Antrea gateway. when a new ClusterIP is created, the
destination IP block of the routing entry might be extended to
include the ClusterIP address

To support the Service traffic of above cases, the main changes of
OVS pipeline include:
- Add table serviceConntrackCommitTable 106 to perform SNAT for
Service traffic.
- Modify table hairpinSNATTable ID from 106 to 108.
- Modify table serviceHairpinTable ID from 29 to 23.
- Add table serviceConntrackTable 24 to transform SNATed connnections.
- Add table serviceClassifierTable 35 to classify the Service traffic.
- Add table serviceConntrackCommitTable 106 to do SNAT for Service traffic.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>