Send command to Suricata socket file directly instead of calling suricatasc
When creating a L7 NetworkPolicy, suricatasc is used to communicate
with Suricata to create/delete something related with the L7
NetworkPolicy. suricatasc is a Python script to send command to the
Suricata socket file. In this PR, command is sent to the Suricata
socket file directly instead of calling suricatasc, which can improve

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>