Fix NodeNetworkPolicy e2e test failure
In NodeNetworkPolicy e2e tests, we have the following cases:

- Node to Node. We deploy two hostNetwork Pods on different Nodes.
- Node to Pods. We deploy a hostNetwork Pod on a Node and two
  non-hostNetwork Pods on different Nodes.

For the case of Node to Pods, after creating test Pods, a full mesh
probing is run to ensure that all Pods can be reachable from each
other. However, the UDP probing from a non-hostNetwork Pod to the
hostNetwork Pod deployed on the same Node will get a failure. The
reason is that due to UDP's connectionless nature, the reply traffic
may use a source IP address determined by routing decisions or outgoing
interfaces, rather than the destination IP address used in request
traffic. This can lead to probing failures for the UDP server.

To resolve the issue, we ensure that the UDP server listens exclusively
on Node IPs when the test Pod is a host network Pod.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>